Expand the popup and status type info to include params for arrow types.
Note that this an improvement on the information available previously, but there
is still room for more. In particular, in the tooltip, only the return type is
hot-linked right now, ensime seems not to give back rich param information at
present so that will need to be added. Made the decision that this is enough
of an improvement for its own PR though.