Cope with no ensime env and deactivate plugin correctly
Previously the plugin was attempting to run on all project windows
including those without any ensime configuration, resulting
in lots of exceptions and interruption of the ensime and other
plugins. Env now detects when there is no ensime environment
or if it is bad (catches the exception) and disables all of the
ensime functionality accordingly for that window/buffer. It works
as expected on projects with ensime files present.