Use window folder for ensime detection, when available
My previous "fix" meant that new windows opened on an ensime project
did not initialize correctly (the ensime detection kicked in before
the folder was set for the window, by that stage the caching had
decided that the window had no ensime and would never try and create
one). Now it uses the folder for the window, so when you switch to
the window, the folder path is used as a key and even if it detected
no ensime env, that will be under another key and not interfere with
this one being detected.