Display an error message if we can't find SBT when starting the server
Ensime Sublime uses SBT to bootstrap the server.
To do this we need SBT to be accessible on the PATH being used by Sublime Text.
Sometimes this PATH is different to the PATH in the user's shell.

Previously, if we couldn't find SBT, Ensime Sublime would simply stop
before calling exec_save_classpath. The last message on the console would be:

    Running sbt saveClasspath (in /path/to/.ensime_cache/Resolution)

Now we display an error message instead.